Implement from for into `MaybeProp<TextProp>`
