+++ leptos_dom/src/macro_helpers/into_attribute.rs
@@ -1,8 +1,8 @@
-use leptos_reactive::Oco;
 #[cfg(not(feature = "nightly"))]
 use leptos_reactive::{
     MaybeProp, MaybeSignal, Memo, ReadSignal, RwSignal, Signal, SignalGet,
 };
+use leptos_reactive::{Oco, TextProp};
 use std::{borrow::Cow, rc::Rc};
 #[cfg(all(target_arch = "wasm32", feature = "web"))]
 use wasm_bindgen::UnwrapThrowExt;
@@ -256,6 +256,14 @@ impl IntoAttribute for Option<Box<dyn IntoAttribute>> {
             Some(bx) => bx.into_attribute_boxed(),
             None => Attribute::Option(None),
         }
+    }
+
+    impl_into_attr_boxed! {}
+}
+
+impl IntoAttribute for TextProp {
+    fn into_attribute(self) -> Attribute {
+        self.get().into_attribute()
     }
 
     impl_into_attr_boxed! {}
